Updated titanium.sh for provisioning builds

+++ bin/titanium.sh
@@ -13,6 +13,10 @@ HOCKEY_ENABLED=${hockey}
 APK_ONLY=${justapk}
 RELEASE_NOTES=${notes}
 IPHONE_DEV_CERT=${cert}
+PROVISIONING_PROFILE_NAME=${profile_file}
+BUILD_ACTION=${action}
+PROFILE_TYPE=iphone_dev_name
+
 # Look all over for a titanium install
 for d in /Users/*
 do
@@ -39,6 +43,18 @@ done
 if [ "${APP_DEVICE}" == "" ]; then
 	APP_DEVICE="iphone"
 fi
+
+# only install|adhoc are supported as install actions
+if [ ! "${BUILD_ACTION}" == "install" ] && [ ! "${BUILD_ACTION}" == "adhoc" ]; then
+	echo ""
+	echo "[WARN] Only action=install and action=adhoc are supported. Choosing action=install."
+	echo ""
+fi
+
+if [ "${BUILD_ACTION}" == "adhoc" ]; then
+	PROFILE_TYPE=iphone_dist_name
+fi
+
 
 # Make sure an SDK version is set
 if [ "${TI_SDK_VERSION}" == "" ]; then
@@ -112,7 +128,7 @@ if [ ${APP_DEVICE} == "iphone" -o ${APP_DEVICE} == "ipad" ]; then
 		bash -c "'${TI_IPHONE_DIR}/prereq.py' package" | \
 		while read prov
 		do
-			temp_iphone_dev_names=`echo $prov | python -c 'import json,sys;obj=json.loads(sys.stdin.read());print obj["'"iphone_dev_name"'"]'| sed 's/ u//g' | sed 's/\[u//g' | sed 's/\[//g'| sed 's/\]//g'| sed "s/\ '//g"| sed "s/\'//g"`
+			temp_iphone_dev_names=`echo $prov | python -c 'import json,sys;obj=json.loads(sys.stdin.read());print obj["'"$PROFILE_TYPE"'"]'| sed 's/ u//g' | sed 's/\[u//g' | sed 's/\[//g'| sed 's/\]//g'| sed "s/\ '//g"| sed "s/\'//g"`
 			IFS=,
 			IPHONE_DEV_NAMES=(${temp_iphone_dev_names//,iphone_dev_name:/})
 
@@ -134,10 +150,10 @@ if [ ${APP_DEVICE} == "iphone" -o ${APP_DEVICE} == "ipad" ]; then
 			fi
 
 			SIGNING_IDENTITY=${IPHONE_DEV_NAMES[$IPHONE_DEV_CERT]}
-			PROVISIONING_PROFILE="${PROJECT_ROOT}/certs/development.mobileprovision"
+			PROVISIONING_PROFILE="${PROJECT_ROOT}/certs/$PROVISIONING_PROFILE_NAME.mobileprovision"
 
-            if [ ! -r 'certs/development.mobileprovision' ];then
-				echo "You must have a file called ${PROVISIONING_PROFILE} to beild for device..."
+            if [ ! -r 'certs/$PROVISIONING_PROFILE_NAME.mobileprovision' ];then
+				echo "You must have a file called ${PROVISIONING_PROFILE} to build for device..."
 				exit
             fi
 
@@ -151,12 +167,20 @@ if [ ${APP_DEVICE} == "iphone" -o ${APP_DEVICE} == "ipad" ]; then
 				temp_array=(${line//{\"uuid\": \"/})
 
 				UUID=${temp_array[0]//\"/}
-                echo "'${TI_IPHONE_BUILD}' install ${iphone} '${PROJECT_ROOT}/' $(echo $APP_ID) '$(echo $APP_NAME)' '$(echo $UUID | sed -e "s/uuid: //g")' '${SIGNING_IDENTITY}' '$(echo ${APP_DEVICE})'"
-				bash -c "'${TI_IPHONE_BUILD}' install ${iphone} '${PROJECT_ROOT}/' $(echo $APP_ID) '$(echo $APP_NAME)' '$(echo $UUID | sed -e "s/uuid: //g")' '${SIGNING_IDENTITY}' '$(echo ${APP_DEVICE})'" | \
+                echo "'${TI_IPHONE_BUILD}' $BUILD_ACTION ${iphone} '${PROJECT_ROOT}/' $(echo $APP_ID) '$(echo $APP_NAME)' '$(echo $UUID | sed -e "s/uuid: //g")' '${SIGNING_IDENTITY}' '$(echo ${APP_DEVICE})'"
+				bash -c "'${TI_IPHONE_BUILD}' $BUILD_ACTION ${iphone} '${PROJECT_ROOT}/' $(echo $APP_ID) '$(echo $APP_NAME)' '$(echo $UUID | sed -e "s/uuid: //g")' '${SIGNING_IDENTITY}' '$(echo ${APP_DEVICE})'" | \
 				while read build_log
 				do
-
-					if [ "${build_log}" == '[INFO] iTunes sync initiated' ]; then
+					MAY_SYNC=0
+					if [ $BUILD_ACTION == "install" ] && [ "${build_log}" == '[INFO] iTunes sync initiated' ]; then
+						MAY_SYNC=1
+						BUILD_LOCATION="Debug-iphoneos"
+					elif [ $BUILD_ACTION == "adhoc" ] && [ "${build_log}" =~ 'PackageApplication' ]; then
+						MAY_SYNC=1
+						BUILD_LOCATION="Release-iphoneos"
+					fi
+					
+					if [ $MAY_SYNC -eq 1 ]; then
 
 						echo "[INFO] Done building app..."\
 						| perl -pe 's/^\[DEBUG\].*$/\e[35m$&\e[0m/g;s/^\[INFO\].*$/\e[36m$&\e[0m/g;s/^\[WARN\].*$/\e[33m$&\e[0m/g;s/^\[ERROR\].*$/\e[31m$&\e[0m/g;'
@@ -183,7 +207,9 @@ if [ ${APP_DEVICE} == "iphone" -o ${APP_DEVICE} == "ipad" ]; then
 							echo "[INFO] Creating .ipa from compiled app"\
 							| perl -pe 's/^\[DEBUG\].*$/\e[35m$&\e[0m/g;s/^\[INFO\].*$/\e[36m$&\e[0m/g;s/^\[WARN\].*$/\e[33m$&\e[0m/g;s/^\[ERROR\].*$/\e[31m$&\e[0m/g;'
 
-							/bin/rm "/tmp/$(echo $APP_NAME).ipa"
+							if [ -f /tmp/$(echo $APP_NAME).ipa ]; then
+								/bin/rm "/tmp/$(echo $APP_NAME).ipa"
+							fi
 							/usr/bin/xcrun -sdk iphoneos PackageApplication -v "${APP}" -o "/tmp/$(echo $APP_NAME).ipa" --sign "${SIGNING_IDENTITY}" --embed "${PROVISIONING_PROFILE}" | \
 							while read package_log
 							do
